+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/autoinstall/spring/Spring7InstallStrategy.kt
@@ -9,7 +9,7 @@ import org.gradle.api.artifacts.dsl.ComponentMetadataHandler
 import org.slf4j.Logger
 
 // @CacheableRule
-abstract class Spring6InstallStrategy : AbstractInstallStrategy {
+abstract class Spring7InstallStrategy : AbstractInstallStrategy {
 
   constructor(logger: Logger) : super() {
     this.logger = logger
@@ -20,7 +20,7 @@ abstract class Spring6InstallStrategy : AbstractInstallStrategy {
   constructor() : this(SentryPlugin.logger)
 
   override val sentryModuleId: String
-    get() = SENTRY_SPRING_6_ID
+    get() = SENTRY_SPRING_7_ID
 
   override val minSupportedThirdPartyVersion: SemVer
     get() = MIN_SUPPORTED_VERSION
@@ -29,18 +29,18 @@ abstract class Spring6InstallStrategy : AbstractInstallStrategy {
     get() = MAX_SUPPORTED_VERSION
 
   override val minSupportedSentryVersion: SemVer
-    get() = SemVer(6, 7, 0)
+    get() = SemVer(8, 21, 0)
 
   companion object Registrar : InstallStrategyRegistrar {
     private const val SPRING_GROUP = "org.springframework"
-    private const val SPRING_6_ID = "spring-core"
-    internal const val SENTRY_SPRING_6_ID = "sentry-spring-jakarta"
+    private const val SPRING_7_ID = "spring-core"
+    internal const val SENTRY_SPRING_7_ID = "sentry-spring-7"
 
-    private val MIN_SUPPORTED_VERSION = SemVer(6, 0, 0)
-    private val MAX_SUPPORTED_VERSION = SemVer(6, 9999, 9999)
+    private val MIN_SUPPORTED_VERSION = SemVer(7, 0, 0, "M1")
+    private val MAX_SUPPORTED_VERSION = SemVer(7, 9999, 9999)
 
     override fun register(component: ComponentMetadataHandler) {
-      component.withModule("$SPRING_GROUP:$SPRING_6_ID", Spring6InstallStrategy::class.java) {}
+      component.withModule("$SPRING_GROUP:$SPRING_7_ID", Spring7InstallStrategy::class.java) {}
     }
   }
 }